refactor: extract useTaskActions hook
Moves all task CRUD operations out of App.jsx into useTaskActions:
- Deadline: setDeadline, postponeDeadlineTask, clearDeadline
- Create: addTask, openNewTaskForm, openNewInboxTask
- Update: changeTaskColor, updateTaskNotes, updateRecurrencePattern,
  updateRecurrenceEndCondition, updateRecurringTemplate
- Complete: toggleComplete
- Move: postponeTask, moveToInbox
- Subtasks: addSubtask, toggleSubtask, deleteSubtask, updateSubtaskTitle
- Delete: moveToRecycleBin, deleteRecurringInstance, recordDeletedTaskTombstone
- Schedule: scheduleTaskAtNextSlot, manuallyScheduleTask
- Focus wrappers: focusCompleteTask, focusUpdateTaskNotes, focusAddSubtask,
  focusToggleSubtask, focusDeleteSubtask, focusUpdateSubtaskTitle

App.jsx reduced by ~620 lines.
